# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation about projekt SasiadWPotrzebie",
        "description": "sasiadwpotrzebie api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "local server"
        },
        {
            "url": "http://localhost:9000",
            "description": "staging server"
        },
        {
            "url": "http://localhost:9000",
            "description": "production server"
        }
    ],
    "paths": {
...
        "/api/aid-categories": {
            "get": {
                "tags": [
                    "Aid Categories"
                ],
                "summary": "Retrieve all aid categories",
                "description": "Get all aidCategories",
                "operationId": "82ce2573ea384e69090603ed0fed8baa",
                "responses": {
                    "200": {
                        "description": "List of all aid categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Emergency Aid"
                                            },
                                            "aid_type_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-22T10:20:30Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-23T15:30:45Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/aid-categories/{id}": {
            "get": {
                "tags": [
                    "Aid Categories"
                ],
                "summary": "Retrieve an aid category by its ID",
                "description": "Get aid category by id",
                "operationId": "da5343b73ed367c2afbe5b44da3114a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Aid category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aid category details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Emergency Aid"
                                        },
                                        "aid_type_id": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aid category not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/aid-categories/{id}/update": {
            "put": {
                "tags": [
                    "Aid Categories"
                ],
                "summary": "Update an aid category by ID",
                "description": "Update aid category by id",
                "operationId": "5977013580ad48353d2f1e34a56a039a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Aid category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Aid Category"
                                    },
                                    "aid_type_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aid category updated successfully"
                    },
                    "404": {
                        "description": "Aid category not found"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/aid-categories/{id}/delete": {
            "delete": {
                "tags": [
                    "Aid Categories"
                ],
                "summary": "Delete an aid category by ID",
                "description": "Delete aid category by id",
                "operationId": "1617c3003bf1d809848ac84abe560399",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Aid category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aid category deleted successfully"
                    },
                    "404": {
                        "description": "Aid category not found"
                    }
                }
            }
        },
        "/api/aid-categories/create": {
            "post": {
                "tags": [
                    "Aid Categories"
                ],
                "summary": "Create a new aid category",
                "description": "Create aid category",
                "operationId": "7aac40c75857e931c176403744a0cdb2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Aid Category"
                                    },
                                    "aid_type_id": {
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Aid category created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}